+++ src/Control/Effect/Error.hs
@@ -34,7 +34,11 @@ throwError = send . Throw
 
 -- | Run a computation which can throw errors with a handler to run on error.
 --
---   Errors thrown by the handler will escape up to the nearest enclosing 'catchError' (if any).
+-- Errors thrown by the handler will escape up to the nearest enclosing 'catchError' (if any).
+-- Note that this effect does /not/ handle errors thrown from impure contexts such as IO,
+-- nor will it handle exceptions thrown from pure code. If you need to handle IO-based errors,
+-- consider if 'Control.Effect.Resource' fits your use case; if not, use 'liftIO' with
+-- 'Control.Exception.try' or use 'Control.Exception.Catch' from outside the effect invocation.
 --
 --   prop> run (runError (pure a `catchError` pure)) == Right a
 --   prop> run (runError (throwError a `catchError` pure)) == Right @Int @Int a
@@ -46,7 +50,7 @@ catchError m h = send (Catch m h ret)
 -- | Run an 'Error' effect, returning uncaught errors in 'Left' and successful computations’ values in 'Right'.
 --
 --   prop> run (runError (pure a)) == Right @Int @Int a
-runError :: (Carrier sig m, Effect sig, Monad m) => Eff (ErrorC exc m) a -> m (Either exc a)
+runError :: forall exc sig m a . (Carrier sig m, Effect sig, Monad m) => Eff (ErrorC exc m) a -> m (Either exc a)
 runError = runErrorC . interpret
 
 newtype ErrorC e m a = ErrorC { runErrorC :: m (Either e a) }